+++ packages/Avatar/src/Avatar.styles.js
@@ -1,47 +1,94 @@
-import { spacer } from "@paprika/stylers/lib/helpers";
+import { fontSize, spacer } from "@paprika/stylers/lib/helpers";
 import styled from "styled-components";
-import stylers from "@paprika/stylers";
 import tokens from "@paprika/tokens";
 import { getAvatarColors } from "./helpers";
 
-const smallSize = `${spacer(4)}`;
-const mediumSize = `${spacer(5)}`;
-
-export const avatarSizeStyles = {
+const roundStyles = {
   small: `
-    border-radius: ${tokens.button.borderRadius};
-    height: ${smallSize};
-    width: ${smallSize};
-    ${stylers.fontSize(2)};
+    border-radius: 50%;
+    height: ${spacer(2.5)};
+    width: ${spacer(2.5)};
   `,
-
   medium: `
-    border-radius: ${tokens.card.borderRadius};
-    height: ${mediumSize};
-    width: ${mediumSize};
-    ${stylers.fontSize(3)};
+    border-radius: 50%;
+    height: ${spacer(3)};
+    width: ${spacer(3)};
+  `,
+  large: `
+    border-radius: 50%;
+    height: ${spacer(4)};
+    width: ${spacer(4)};
   `,
 };
+
+const roundStringStyles = {
+  small: fontSize(-4),
+  medium: "font-size: 12px;",
+  large: fontSize(-1),
+};
+
+const roundIconStyles = {
+  small: fontSize(-2),
+  medium: fontSize(-1),
+  large: fontSize(2),
+};
+
+const squareStyles = {
+  medium: `
+    border-radius: ${tokens.button.borderRadius};
+    height: ${spacer(4)};
+    width: ${spacer(4)};
+  `,
+
+  large: `
+    border-radius: ${tokens.card.borderRadius};
+    height: ${spacer(5)};
+    width: ${spacer(5)};
+  `,
+};
+
+const squareStringStyles = {
+  medium: fontSize(0),
+  large: fontSize(2),
+};
+
+const squareIconStyles = {
+  medium: fontSize(2),
+  large: fontSize(4),
+};
+
+function getSizeCss(size, isRound, isChildString) {
+  let sizeValue;
+
+  if (isRound) {
+    sizeValue = roundStyles[size].concat(isChildString ? roundStringStyles[size] : roundIconStyles[size]);
+  } else {
+    sizeValue = squareStyles[size].concat(isChildString ? squareStringStyles[size] : squareIconStyles[size]);
+  }
+
+  return sizeValue;
+}
 
 export const Avatar = styled.div`
   align-items: center;
   display: inline-flex;
   font-weight: bold;
   justify-content: center;
+  overflow: hidden;
 
   &,
   * {
     box-sizing: border-box;
   }
 
-  ${({ $backgroundColor, $color, size, children }) => {
-    const sizeValue = avatarSizeStyles[size];
-    const color =
-      typeof children !== "string"
-        ? { backgroundColor: tokens.color.blackLighten60, fontColor: tokens.color.blackLighten20 }
-        : getAvatarColors(children);
+  ${({ $backgroundColor, $color, isChildString, isRound, size, children }) => {
+    const color = isChildString
+      ? getAvatarColors(children)
+      : { backgroundColor: tokens.color.blackLighten60, fontColor: tokens.color.blackLighten20 };
     const backgroundColor = $backgroundColor ?? color.backgroundColor;
     const fontColor = $color ?? color.fontColor;
+    const sizeValue = getSizeCss(size, isRound, isChildString);
+
     return `
       background-color: ${backgroundColor};
       color: ${fontColor};