+++ packages/FormElement/src/components/Layout/Layout.js
@@ -4,7 +4,13 @@ import * as sc from "./Layout.styles";
 import LeftCol from "./LeftCol/LeftCol";
 import RightCol from "./RightCol/RightCol";
 
+function Layout(props) {
+  const { children, ...moreProps } = props;
+  return <sc.Layout {...moreProps}>{children}</sc.Layout>;
+}
+
 const propTypes = {
+  /** Content of the flex parent */
   children: PropTypes.node,
 };
 
@@ -12,16 +18,9 @@ const defaultProps = {
   children: null,
 };
 
-function Layout(props) {
-  const { children, ...moreProps } = props;
-  return <sc.Layout {...moreProps}>{children}</sc.Layout>;
-}
-
 Layout.displayName = "FormElement.Layout";
-
 Layout.propTypes = propTypes;
 Layout.defaultProps = defaultProps;
-
 Layout.LeftCol = LeftCol;
 Layout.RightCol = RightCol;