+++ packages/ListBox/stories/WithTags/examples/Sizes.js
@@ -39,7 +39,6 @@ function App({ size = "medium", isOpen = false }) {
 export default function Sizes() {
   return (
     <>
-      <App size="small" />
       <App size="medium" isOpen />
       <br />
       <br />