+++ packages/Avatar/CHANGELOG.md
@@ -17,3 +17,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Changed the border-radius of the small and the medium avatar.
+
+## [2.0.0] - 2020-12-07
+
+### Changed
+
+- Breaking change: The old "MEDIUM" square Avatar is now "LARGE". The old "SMALL" square Avatar is now "MEDIUM".
+- Breaking change: the Avatar will display whatever you pass into it as-is (including capitalization). Created the `getInitialsFromText` helper function so the consumer can get the text they want to pass in.
+- Added new "isRound" prop so you can have round Avatars in three sizes