+++ packages/Popover/CHANGELOG.md
@@ -20,3 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Updated: Created and placed all constants into type.js file. [@kaan.darcey](https://github.com/KDarcey).
+
+## [1.0.4-alpha.0] - 2020-12-16
+
+- Pass the `Event` object down for `<Popover.Trigger />` `onClick`.